requirements.txt
boto3==1.13.11
botocore==1.16.11
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
Pillow==7.1.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
twilio==6.40.0
urllib3==1.25.9
Werkzeug==1.0.1
